requirements.txt
ot
numpy>=1.23
faiss-cpu>=1.7.4
lapjv>=1.3
matplotlib>=3.7
